+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Completion for commands, `:`, using `nvim-cmp` and `cmp-cmdline` and
   `cmp-path` as sources.
 
+- Completion for search, `/`, using `nvim-cmp` and `cmp-cmdline` and
+  `cmd-buffer` as sources.
+
+### Added
+
 - Add completion for word present in current buffers using `nvim-cmp` and
   `cmp-buffer` plugins